% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cpes}
\alias{cpes}
\title{Common Platform Enumeration Entries}
\format{A data frame with +100,000 rows and 3 variables:
\describe{
  \item{cpe.22}{Title of the platform}
  \item{cpe.23}{Identifier for given platform}
  \item{cpe.ref}{URLs encoded as json with additional information regarding the cpe entry}
  ...
}}
\source{
\url{https://cpe.mitre.org}
}
\usage{
cpes
}
\description{
A dataset containing entries with the information of the different platforms containing
over  100,000 rows.
}
\keyword{datasets}

