% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cves.nist}
\alias{cves.nist}
\title{Common Vulnerabilities and Exposures from NIST}
\format{A data frame with 77,630 rows and 19 variables
\describe{
  \item{osvdb.ext}{}
  \item{vulnerable.configuration}{}
  \item{vulnerable.software.list}{}
  \item{cve.id}{}
  \item{discovered.datetime}{}
  \item{disclosure.datetime}{}
  \item{exploit.publish.datetime}{}
  \item{published.datetime}{}
  \item{last.modified.datetime}{}
  \item{cvss}{}
  \item{security.protection}{}
  \item{assessment.check}{}
  \item{cwe}{}
  \item{references}{}
  \item{fix.action}{}
  \item{scanner}{}
  \item{summary}{}
  \item{technical.description}{}
  \item{attack.scenario}{}
}}
\source{
\url{https://nvd.nist.gov/download.cfm}
}
\usage{
cves.nist
}
\description{
A dataset containing the entries for the details of 77,630 CVE entries obtained from NIST
publicly available database.
}
\keyword{datasets}

