% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cves}
\alias{cves}
\title{Common Vulnerabilities and Exposures}
\format{A data frame with 53940 rows and 10 variables:
\describe{
 \item{cve}{CVE Identifier.}
 \item{status}{Factor, The status of the entry, either \emph{Candidate} of \emph{Entry}.}
 \item{description}{Brief description of the entry.}
 \item{references}{Additional infomation regarding the CVE entry.}
 \item{phase}{Phase of entry, including date of assignation}
 \item{votes}{Information regarding the votes of the different reviewers}
 \item{comments}{Additional notes regarding the entry.}
}}
\source{
\url{https://cve.mitre.org}
}
\usage{
cves
}
\description{
A dataset containing the entries for the details for more than 92,000 CVE entries.
}
\keyword{datasets}

