% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cwes}
\alias{cwes}
\title{Common Weakness Enumeration}
\format{A data frame with 719 rows and 26 variables:
\describe{
 \item{ID}{CWE Identifier}
 \item{Name}{}
 \item{Weakness_Abstraction}{}
 \item{Status}{}
 \item{code_standard}{}
 \item{descr.summary}{}
 \item{descr.details}{}
 \item{relationships}{}
 \item{cwe.parents}{}
 \item{ordinalities}{}
 \item{platforms}{}
 \item{time.intro}{}
 \item{consequences}{}
 \item{mitigation}{}
 \item{causal}{}
 \item{demos}{}
 \item{mapping}{}
 \item{history}{}
 \item{relationship.notes}{}
 \item{maintenance.notes}{}
 \item{background}{}
 \item{introduction.mode}{}
 \item{other.notes}{}
 \item{aff.resources}{}
 \item{exploits}{}
 \item{functional.areas}{}
}}
\source{
\url{https://cwe.mitre.org}
}
\usage{
cwes
}
\description{
A dataset containing the entries for the details for the 719 CWE entries.
}
\keyword{datasets}

