% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{DataSetAvailable}
\alias{DataSetAvailable}
\title{DataSetAvailable}
\usage{
DataSetAvailable(dataset = "cves")
}
\arguments{
\item{dataset}{Selects the data set for this operation. Default set to "all". Check available option with DataSetList()}
}
\value{
TRUE if dataset is available, FALSE if it needs an update.
}
\description{
DataSetAvailable
}
