% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{DataSetList}
\alias{DataSetList}
\title{Show data set status.}
\usage{
DataSetList()
}
\value{
List of available datasets.
}
\description{
\code{DataSetList} Show data set status. Prints information about update status and number of observations of local data sets.
}
\details{
Check the internal data structure and returns a character vector with names of available data.frames.
}
\examples{
net.security::DataSetList()
}
