% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{DataSetUpdate}
\alias{DataSetUpdate}
\title{DataSetUpdate}
\usage{
DataSetUpdate(ds = "all")
}
\arguments{
\item{ds}{Selects the data set for this operation. Default set to "all". Check available option with DataSetList()}
}
\value{
Official source files download date time.
}
\description{
DataSetUpdate
}
\examples{
\dontrun{
net.security::DataSetUpdate(ds = "all")
}
\dontrun{
net.security::DataSetUpdate(ds = "cves")
}
}
