% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cve.R
\name{DownloadCVEData}
\alias{DownloadCVEData}
\title{DownloadCVEData}
\usage{
DownloadCVEData(dest)
}
\arguments{
\item{dest}{String with directory where to store files to be downloaded.}
}
\description{
DownloadCVEData
}

