% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwe.R
\name{DownloadCWEData}
\alias{DownloadCWEData}
\title{DownloadCWEData, it downloads https://cwe.mitre.org/data/xml/views/2000.xml.zip
that containts all kind of information related with CWE}
\usage{
DownloadCWEData(path = "inst/tmpdata")
}
\description{
DownloadCWEData, it downloads https://cwe.mitre.org/data/xml/views/2000.xml.zip
that containts all kind of information related with CWE
}

