% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ripe.R
\name{DownloadRIPE.ASN}
\alias{DownloadRIPE.ASN}
\title{DownloadRIPE.ASN}
\usage{
DownloadRIPE.ASN(path = "inst/tmpdata")
}
\arguments{
\item{path}{where files will be stored}
}
\value{
path of asn file unziped
}
\description{
DownloadRIPE.ASN
}

