% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ripe.R
\name{DownloadRIPE.inet}
\alias{DownloadRIPE.inet}
\title{DownloadRIPE.inet}
\usage{
DownloadRIPE.inet(path = "inst/tmpdata")
}
\arguments{
\item{path}{where files will be stored}
}
\value{
path of asn file unziped
}
\description{
DownloadRIPE.inet
}
\examples{
path <- DownloadRIPE.inet()
}

