% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cve.R
\name{ExtractCVEFiles}
\alias{ExtractCVEFiles}
\title{ExtractCVEFiles, Extract compressed files}
\usage{
ExtractCVEFiles(path)
}
\arguments{
\item{path}{character, the directory containing the files to be extracted}
}
\description{
ExtractCVEFiles, Extract compressed files
}
