% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwe.R
\name{ExtractCWEFiles}
\alias{ExtractCWEFiles}
\title{Extract compressed files}
\usage{
ExtractCWEFiles(savepath)
}
\arguments{
\item{savepath}{String, the directory containing the files to be extracted}
}
\description{
Extract compressed files
}
