% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cve.R
\name{ExtractDataFiles}
\alias{ExtractDataFiles}
\title{Extract compressed files}
\usage{
ExtractDataFiles(path)
}
\arguments{
\item{path}{String, the directory containing the files to be extracted}
}
\description{
Extract compressed files
}

