% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capec.R
\name{GetCAPECData}
\alias{GetCAPECData}
\title{Download, parse and return a tidy data frame with CAPEC information}
\usage{
GetCAPECData(savepath = tempdir())
}
\arguments{
\item{savepath}{String}
}
\value{
data frame
}
\description{
Download, parse and return a tidy data frame with CAPEC information
}
