% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cve.R
\name{GetCVEData}
\alias{GetCVEData}
\title{Get data frame with CVE information}
\usage{
GetCVEData(origin = "all", savepath = tempdir())
}
\arguments{
\item{origin}{"all", "mitre", "nist"}

\item{savepath}{String}
}
\value{
Data frame
}
\description{
Get data frame with CVE information
}
