% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwe.R
\name{GetCWEChildrenIDs}
\alias{GetCWEChildrenIDs}
\title{GetCWEChildrenIDs, return all CWEs direct child of input CWE}
\usage{
GetCWEChildrenIDs(cwe = "102")
}
\arguments{
\item{cwe, }{only one. Default="102"}
}
\value{
character
}
\description{
GetCWEChildrenIDs, return all CWEs direct child of input CWE
}
\examples{
GetCWEChildrenIDs()
GetCWEChildrenIDs("675")
}

