% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwe.R
\name{GetCWEData}
\alias{GetCWEData}
\title{Get data frame with CWE information}
\usage{
GetCWEData(savepath = tempdir())
}
\arguments{
\item{savepath}{String}
}
\value{
data frame
}
\description{
Get data frame with CWE information
}
