% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{GetCWEHierarcy}
\alias{GetCWEHierarcy}
\title{Given a net.security CWES data.frame it returns a data.frame prepared for
network representation of CWE relationships as edges.}
\usage{
GetCWEHierarcy(as_numbers = T)
}
\arguments{
\item{as_numbers}{if TRUE src and target are numbers, if FALSE as character starting with "CWE-"}

\item{cwes}{data.frame, from net.security data sets}
}
\value{
data.frame
}
\description{
Given a net.security CWES data.frame it returns a data.frame prepared for
network representation of CWE relationships as edges.
}
