% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwe.R
\name{GetCWETitle}
\alias{GetCWETitle}
\title{GetCWETitle, returns CWE's title}
\usage{
GetCWETitle(cwe = "100")
}
\value{
character
}
\description{
GetCWETitle, returns CWE's title
}
\examples{
GetCWETitle(cwe = "99")
}

