% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{GetDataFrame}
\alias{GetDataFrame}
\title{GetDataFrame}
\usage{
GetDataFrame(dataset = "cves")
}
\arguments{
\item{dataset}{Selects the data set for this operation. Default set to "all". Check available option with DataSetList()}
}
\value{
Selected dataset as tidy data.frame
}
\description{
GetDataFrame
}
\examples{
cves <- net.security::GetDataFrame(dataset = "cves")
}
