% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inventary.R
\name{GetInventary}
\alias{GetInventary}
\title{Return data.frame with installed software name, version and vendor.}
\usage{
GetInventary()
}
\description{
Return data.frame with installed software name, version and vendor.
}
