% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cve.R
\name{GetNISTEntry}
\alias{GetNISTEntry}
\title{Parse from NIST XML CVE entry to data.frame}
\usage{
GetNISTEntry(node)
}
\arguments{
\item{node}{XML Node}
}
\value{
data frame
}
\description{
Parse from NIST XML CVE entry to data.frame
}

