% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cve.R
\name{GetNISTvulns}
\alias{GetNISTvulns}
\title{GetNISTvulns}
\usage{
GetNISTvulns()
}
\value{
data frame
}
\description{
> system.time({cve.nist <- GetNISTvulns()})
user  system elapsed
394.63    8.24  435.69
}
\examples{
cve.nist <- GetNISTvulns()
}

