% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwe.R
\name{GetParents}
\alias{GetParents}
\title{Get parents from a given CWE
Given a CWE code, returns its direct parents.set compact=T and results will be semicolon-separated}
\usage{
GetParents(cwes, CWE = "", compact = FALSE)
}
\arguments{
\item{cwes}{data frame}

\item{CWE}{number}

\item{compact}{boolean}
}
\description{
Get parents from a given CWE
Given a CWE code, returns its direct parents.set compact=T and results will be semicolon-separated
}

