% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ripe.R
\name{GetRIPE.ASN}
\alias{GetRIPE.ASN}
\title{GetRIPE.ASN}
\usage{
GetRIPE.ASN(path = "inst/tmpdata", download = FALSE)
}
\arguments{
\item{path}{where RIPE files will be downloaded and unziped (don't finish with /). Default set as inst/tmpdata}

\item{download}{TRUE if you want to update the source data, default set as FALSE}
}
\value{
data frame
}
\description{
GetRIPE.ASN
}
\examples{
ripe.asn <- GetRIPE.ASN()
ripe.asn <- GetRIPE.ASN(download = TRUE)
}

