% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ripe.R
\name{GetRIPE.inet}
\alias{GetRIPE.inet}
\title{GetRIPE.inet}
\usage{
GetRIPE.inet(path = "inst/tmpdata", download = FALSE, use.cached = FALSE)
}
\arguments{
\item{path}{where RIPE files will be downloaded and unziped (don't finish with /). Default set as inst/tmpdata}

\item{download}{TRUE if you want to update the source data, default set as FALSE}

\item{use.cached}{if you want to load from internal rds file}
}
\value{
data frame
}
\description{
GetRIPE.inet
}
\examples{
ripe.inet <- GetRIPE.inet()
ripe.inet <- GetRIPE.inet(download = TRUE)
}

