% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{GetSecurityNetwork}
\alias{GetSecurityNetwork}
\title{It builds a data frame with relations between CWE, CVE and CAPEC.}
\usage{
GetSecurityNetwork(as.igraph = F)
}
\arguments{
\item{as.igraph}{if true it will return igraph object, if false a data frame}
}
\value{
data.frame
}
\description{
It builds a data frame with relations between CWE, CVE and CAPEC.
}
