% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvedetails.R
\name{GetSummaryByCVE}
\alias{GetSummaryByCVE}
\title{GetSummaryByCVE returns a string with the description of the vulneravility and language}
\usage{
GetSummaryByCVE(cve = "CVE-2016-0002", lang = "EN")
}
\arguments{
\item{String}{with CVE code in standrad mode (CVE-YYYY-XXXX)}

\item{It's}{implemented for EN(glish) or ES(pañol)}
}
\description{
GetSummaryByCVE returns a string with the description of the vulneravility and language
}
\examples{
description <- GetSummaryByCVE(cve = "CVE-2015-0002", lang = "ES")
}

