% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asn.R
\name{InfoASN}
\alias{InfoASN}
\title{InfoASN based on public information provided by http://bgp.he.net/
Don't abuse, this is just for test purposes.}
\usage{
InfoASN(asn)
}
\arguments{
\item{asn}{number}
}
\description{
InfoASN based on public information provided by http://bgp.he.net/
Don't abuse, this is just for test purposes.
}

