% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cve.R
\name{LastDownloadCVEDate}
\alias{LastDownloadCVEDate}
\title{LastDownloadDate}
\usage{
LastDownloadCVEDate()
}
\value{
character with last official update date YYYY-MM-DD
}
\description{
LastDownloadDate
}
\examples{
last <- LastDownloadCVEDate()
}
