% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitre_cti.R
\name{MapCommonPropierties}
\alias{MapCommonPropierties}
\title{Extract common propierties from attack pattern object (parsed with RJSONIO::fromJSON)}
\usage{
MapCommonPropierties(attack.obj = NA, domain = NA)
}
\arguments{
\item{domain}{must be "pre-attack", "enterprise-attack" or "mobile-attack"}

\item{attack.pattern}{list}
}
\value{
data.frame compliant with CTI USAGE document
}
\description{
Extract common propierties from attack pattern object (parsed with RJSONIO::fromJSON)
}
\examples{
\dontrun{
sf <- "https://github.com/mitre/cti/raw/master/<domain>/<object>/<file>.json"
attack.pattern <- RJSONIO::fromJSON(sf)
df.common <- MapCommonPropierties(attack.pattern)
}
}
