% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitre_cti.R
\name{parseAttck.Mitigation}
\alias{parseAttck.Mitigation}
\title{Read MITRE CTI Repository browsing domain directories to extract data from course-of-action files,
build model and return tidy data.frame with Mitigation variables.}
\usage{
parseAttck.Mitigation()
}
\value{
data.frame
}
\description{
Read MITRE CTI Repository browsing domain directories to extract data from course-of-action files,
build model and return tidy data.frame with Mitigation variables.
}
\examples{
\dontrun{
df.mitigations <- parseAttck.Mitigation()
}
}
