% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitre_cti.R
\name{parseAttck.Techniques}
\alias{parseAttck.Techniques}
\title{Read MITRE CTI Repository browsing domain directories to extract data from attack-pattern files,
map variables from STIX to ATT&CK model and return tidy data.frame with technique variables.}
\usage{
parseAttck.Techniques()
}
\value{
data.frame
}
\description{
Read MITRE CTI Repository browsing domain directories to extract data from attack-pattern files,
map variables from STIX to ATT&CK model and return tidy data.frame with technique variables.
}
\examples{
\dontrun{
df.techniques <- parseAttck.Techniques()
}
}
