% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitre_cti.R
\name{parseAttckENTmodel.tech}
\alias{parseAttckENTmodel.tech}
\title{Read MITRE CTI Repository files in enterprise-attack directory, extract data,
map variables from STIX to ATT&CK model and return tidy data.frame.}
\usage{
parseAttckENTmodel.tech()
}
\value{
data.frame
}
\description{
Read MITRE CTI Repository files in enterprise-attack directory, extract data,
map variables from STIX to ATT&CK model and return tidy data.frame.
}
\examples{
\dontrun{
df.ent <- parseAttckENTmodel.tech()
}
}
