% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitre_cti.R
\name{parseAttckMOBmodel.tech}
\alias{parseAttckMOBmodel.tech}
\title{Read MITRE CTI Repository files in mobile-attack directory, extract data,
map variables from STIX to ATT&CK model and return tidy data.frame.}
\usage{
parseAttckMOBmodel.tech()
}
\value{
data.frame
}
\description{
Read MITRE CTI Repository files in mobile-attack directory, extract data,
map variables from STIX to ATT&CK model and return tidy data.frame.
}
\examples{
\dontrun{
df.mob <- parseAttckMOBmodel.tech()
}
}
