% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitre_cti.R
\name{parseAttckmodel.Groups}
\alias{parseAttckmodel.Groups}
\title{Read MITRE CTI Repository browsing domain directories to extract data from intrusion-set files,
map variables from STIX to ATT&CK model and return tidy data.frame with Group variables.}
\usage{
parseAttckmodel.Groups()
}
\value{
data.frame
}
\description{
Read MITRE CTI Repository browsing domain directories to extract data from intrusion-set files,
map variables from STIX to ATT&CK model and return tidy data.frame with Group variables.
}
\examples{
\dontrun{
df.groups <- parseAttckmodel.Groups()
}
}
