% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitre_cti.R
\name{parseAttckmodel.rels}
\alias{parseAttckmodel.rels}
\title{Title}
\usage{
parseAttckmodel.rels(domain = sample(c("pre-attack", "enterprise-attack",
  "mobile-attack"), 1))
}
\arguments{
\item{domain}{}
}
\value{
data.frame
}
\description{
Title
}
