% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitre_cti.R
\name{parseAttckmodel.tact}
\alias{parseAttckmodel.tact}
\title{Read MITRE CTI Repository files retaled to x-mitre-tactic, extract data,
map variables from STIX to ATT&CK model and return tidy data.frame.}
\usage{
parseAttckmodel.tact(domain = sample(c("pre-attack", "enterprise-attack",
  "mobile-attack"), 1))
}
\arguments{
\item{domain}{must be "pre-attack", "enterprise-attack" or "mobile-attack"}
}
\value{
data.frame
}
\description{
Read MITRE CTI Repository files retaled to x-mitre-tactic, extract data,
map variables from STIX to ATT&CK model and return tidy data.frame.
}
