% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitre_cti.R
\name{parseAttckPREmodel}
\alias{parseAttckPREmodel}
\title{Read MITRE CTI Repository files in pre-attack directory, extract data,
map variables from STIX to ATT&CK model and return tidy data.frame.}
\usage{
parseAttckPREmodel()
}
\value{
data.frame
}
\description{
Read MITRE CTI Repository files in pre-attack directory, extract data,
map variables from STIX to ATT&CK model and return tidy data.frame.
}
\examples{
\dontrun{
df.pre <- parseAttckPREmodel()
}
}
