% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capec.R
\name{ParseCAPECData.categories}
\alias{ParseCAPECData.categories}
\title{Title}
\usage{
ParseCAPECData.categories(doc)
}
\arguments{
\item{cwe.file}{String}
}
\value{
Data frame
}
\description{
Title
}

