% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cve.R
\name{ParseCVEData}
\alias{ParseCVEData}
\title{Arrange CVE information into data frame}
\usage{
ParseCVEData(cve.file)
}
\arguments{
\item{cve.file}{String}
}
\value{
Data frame
}
\description{
Arrange CVE information into data frame
}

