% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cve.R
\name{ParseCVEMITREData}
\alias{ParseCVEMITREData}
\title{Arrange CVE information into data frame}
\usage{
ParseCVEMITREData(path)
}
\arguments{
\item{cve.file}{String}
}
\value{
Data frame
}
\description{
Arrange CVE information into data frame
}

