% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwe.R
\name{ParseCWEData}
\alias{ParseCWEData}
\title{Arrange CWE information into data frame}
\usage{
ParseCWEData(cwe.file)
}
\arguments{
\item{cwe.file}{String}
}
\value{
Data frame
}
\description{
Arrange CWE information into data frame
}

