% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmap.R
\name{ParseNMAP}
\alias{ParseNMAP}
\title{ParseNMAP Parser for nmap xml output}
\usage{
ParseNMAP(file.input, file.output.header = "")
}
\arguments{
\item{file.input}{XML with nmap results}

\item{file.output.header}{string as header for output file}
}
\value{
data frame
}
\description{
ParseNMAP Parser for nmap xml output
}
\examples{
df <- ParseNMAP("nmap.xml","my-")
}

