% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitre_cti.R
\name{ppparseAttckmodel.soft}
\alias{ppparseAttckmodel.soft}
\title{Read MITRE CTI Repository files retaled to malware and tool, extract data,
map variables from STIX to ATT&CK model and return tidy data.frame.}
\usage{
ppparseAttckmodel.soft(domain = sample(c("pre-attack",
  "enterprise-attack", "mobile-attack"), 1))
}
\arguments{
\item{domain}{must be "pre-attack", "enterprise-attack" or "mobile-attack"}
}
\value{
data.frame
}
\description{
Read MITRE CTI Repository files retaled to malware and tool, extract data,
map variables from STIX to ATT&CK model and return tidy data.frame.
}
