% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitre_cti.R
\name{preMapCommonPropierties}
\alias{preMapCommonPropierties}
\title{Extract common propiertis from attack pattern object (parsed with RJSONIO::fromJSON)}
\usage{
preMapCommonPropierties(attack.pattern)
}
\arguments{
\item{attack.pattern}{list}
}
\value{
data.frame compliant with CTI USAGE document
}
\description{
Extract common propiertis from attack pattern object (parsed with RJSONIO::fromJSON)
}
\examples{
\dontrun{
sf <- "https://github.com/mitre/cti/raw/master/<domain>/<object>/<file>.json"
attack.pattern <- RJSONIO::fromJSON(sf)
df.common <- MapCommonPropierties(attack.pattern)
}
}
