% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cve.R
\name{UnzipDataFiles}
\alias{UnzipDataFiles}
\title{UnzipDataFiles}
\usage{
UnzipDataFiles(path)
}
\arguments{
\item{path}{String, the directory containing the files to be extracted}
}
\description{
UnzipDataFiles
}

