% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cve.R
\name{UpdateCVEData}
\alias{UpdateCVEData}
\title{UpdateCVEData download CVE information from public databases such as MITRE and NIST}
\usage{
UpdateCVEData(all = FALSE, lang = "EN")
}
\arguments{
\item{Boolean}{value, if true it will download all data base, if false it will only download last updates.}

\item{It's}{implemented for EN(glish) or ES(pañol)}
}
\description{
UpdateCVEData download CVE information from public databases such as MITRE and NIST
}
\examples{
UpdateCVEData(all = TRUE, lang = "ES")
}

