% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capec.R
\name{XMLChildren2JSON}
\alias{XMLChildren2JSON}
\title{Title}
\usage{
XMLChildren2JSON(doc, xpath.root, root.id, xpath.children, json = TRUE)
}
\arguments{
\item{xpath.children}{}
}
<<<<<<< HEAD
\value{

}
=======
>>>>>>> capec
\description{
Title
}

