% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{DataSetUpdate}
\alias{DataSetUpdate}
\title{DataSetUpdate}
\usage{
DataSetUpdate(dataset = "all")
}
\arguments{
\item{dataset}{Selects the data set for this operation. Default set to "all". Check available option with DataSetList()}
}
\description{
DataSetUpdate
}
