% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cve.R
\name{GetCVEData}
\alias{GetCVEData}
\title{GetCVEData}
\usage{
GetCVEData(path = "inst/tmpdata", download = FALSE)
}
\arguments{
\item{path}{where Standard CVE definitions will be downloaded and unziped (don't finish with /). Default set as inst/tmpdata}

\item{download}{}
}
\value{
data frame
}
\description{
GetCVEData
}
\examples{
cves <- GetCVEData()
cves <- GetCVEData(download = TRUE)
}

