% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwe.R
\name{GetCWEData}
\alias{GetCWEData}
\title{GetCWEData, parse xml.path}
\usage{
GetCWEData(xml.path = "inst/tmpdata/2000.xml")
}
\arguments{
\item{xml.path, }{absolute path. Default = "inst/tmpdata/2000.xml"}
}
\value{
"XMLInternalDocument" "XMLAbstractDocument"
}
\description{
GetCWEData, parse xml.path
}
\examples{
doc <- GetCWEData()
}

