% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwe.R
\name{GetCWEData}
\alias{GetCWEData}
\title{GetCWEData Download current CWE definitions, parse them and return the info as data frame}
\usage{
GetCWEData(path = "inst/tmpdata")
}
\arguments{
\item{path}{where MITRE CWE definitions will be downloaded and unziped (don't finish with /). Default set as inst/tmpdata}
}
\value{
data frame
}
\description{
GetCWEData Download current CWE definitions, parse them and return the info as data frame
}
\examples{
df <- GetCWEData()
df <- GetCWEData("/tmp")
}

