% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwe.R
\name{GetCWEData}
\alias{GetCWEData}
\title{GetCWEData Download current CWE definitions, parse them and return the info as data frame}
\usage{
GetCWEData()
}
\value{
data frame
}
\description{
GetCWEData Download current CWE definitions, parse them and return the info as data frame
}

