% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{GetDataFrame}
\alias{GetDataFrame}
\title{Get data sets as data frames.}
\usage{
GetDataFrame(ds)
}
\arguments{
\item{ds}{Selects the data set for this operation. Default set to "all". Check available option with DataSetList()}
}
\value{
Selected dataset as tidy data.frame
}
\description{
\code{GetDataFrame} Get data sets as data frames. Check data sets documentation for details of data frames.
}
\details{
Returns the data.frame selected. "Unknown" if it's not available.
}
\examples{
cves <- net.security::GetDataFrame(ds = "cves")
}
