% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cve.R
\name{GetNISTEntry}
\alias{GetNISTEntry}
\title{Get CVE entry from downloaded NIST information}
\usage{
GetNISTEntry(node)
}
\arguments{
\item{node}{XML Node}
}
\value{
data frame
}
\description{
Get CVE entry from downloaded NIST information
}

