% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cve.R
\name{GetNISTVulns}
\alias{GetNISTVulns}
\title{GetNISTVulns}
\usage{
GetNISTVulns(years = as.integer(format(Sys.Date(), "\%Y")))
}
\arguments{
\item{years}{numeric vector with values between 2002 and current year}
}
\value{
data frame
}
\description{
GetNISTVulns
}

