% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cve.R
\name{GetNISTvulnsByYear}
\alias{GetNISTvulnsByYear}
\title{Create CVE data.frame from NIST entries for specified year}
\usage{
GetNISTvulnsByYear(year = as.integer(format(Sys.Date(), "\%Y")))
}
\arguments{
\item{year}{value between 2002 and current year, default value is set as current year}
}
\value{
data frame
}
\description{
Create CVE data.frame from NIST entries for specified year
}

