% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwe.R
\name{GetParents}
\alias{GetParents}
\title{Given a CWE code it returns its direct parents, set compact=T and results will be dotcomma-separated}
\usage{
GetParents(cwes, CWE = "", compact = FALSE)
}
\description{
Given a CWE code it returns its direct parents, set compact=T and results will be dotcomma-separated
}

