% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cve.R
\name{UpdateNIST}
\alias{UpdateNIST}
\title{UpdateNIST Update CVE DB from NIST
Reference: https://nvd.nist.gov/download.cfm}
\usage{
UpdateNIST(all = FALSE, spanish = FALSE)
}
\description{
UpdateNIST Update CVE DB from NIST
Reference: https://nvd.nist.gov/download.cfm
}

